requirements.txt
requests
python-memcached
